+++ app/Http/Controllers/ItemCategoryController.php
@@ -51,7 +51,7 @@ class ItemCategoryController extends Controller
         $cache_collection = new Cache\Collection();
         $cache_collection->setFromCache($cache_control->get(request()->getRequestUri()));
 
-        if ($cache_collection->valid() === false) {
+        if ($cache_control->cacheable() === false || $cache_collection->valid() === false) {
 
             $item_category = (new ItemCategory())->paginatedCollection(
                 $resource_type_id,