+++ app/Http/Controllers/PermittedUserController.php
@@ -54,7 +54,7 @@ class PermittedUserController extends Controller
         $cache_collection = new Cache\Collection();
         $cache_collection->setFromCache($cache_control->get(request()->getRequestUri()));
 
-        if ($cache_collection->valid() === false) {
+        if ($cache_control->cacheable() === false || $cache_collection->valid() === false) {
             $total = (new PermittedUser())->totalCount(
                 $resource_type_id,
                 $search_parameters