+++ app/Models/Item/Summary/SimpleExpense.php
@@ -41,7 +41,9 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 category.id, 
                 category.name AS name, 
                 category.description AS description,
-                SUM({$this->sub_table}.total) AS total")->
+                SUM({$this->sub_table}.total) AS total, 
+                COUNT({$this->sub_table}.item_id) AS total_count, 
+                MAX({$this->sub_table}.created_at) AS last_updated")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->
@@ -129,7 +131,8 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
         $collection = $this->
             selectRaw("
                 SUM({$this->sub_table}.total) AS total, 
-                COUNT({$this->sub_table}.item_id) AS total_count
+                COUNT({$this->sub_table}.item_id) AS total_count,
+                MAX({$this->sub_table}.created_at) AS last_updated
             ")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
@@ -185,7 +188,9 @@ class SimpleExpense extends Model implements ISummaryModel, ISummaryModelCategor
                 sub_category.id, 
                 sub_category.name AS name, 
                 sub_category.description AS description,
-                SUM({$this->sub_table}.total) AS total")->
+                SUM({$this->sub_table}.total) AS total, 
+                COUNT({$this->sub_table}.item_id) AS total_count, 
+                MAX({$this->sub_table}.created_at) AS last_updated")->
             join($this->sub_table, 'item.id', "{$this->sub_table}.item_id")->
             join("resource", "resource.id", "item.resource_id")->
             join("resource_type", "resource_type.id", "resource.resource_type_id")->