+++ app/Response/Header/Headers.php
@@ -100,6 +100,20 @@ class Headers
     }
 
     /**
+     * Add the X-Total-Count header
+     *
+     * @param int $total
+     *
+     * @return Headers
+     */
+    public function addTotalCount(int $total): Headers
+    {
+        $this->headers->add('X-Total-Count', $total);
+
+        return $this;
+    }
+
+    /**
      * Add the X-Filter header if the parameters for filtering exist
      *
      * @param string|null $filter_header
@@ -111,6 +125,20 @@ class Headers
         if ($filter_header !== null) {
             $this->headers->addFilter($filter_header);
         }
+
+        return $this;
+    }
+
+    /**
+     * Add the X-Last-Updated header
+     *
+     * @param string|null $last_updated
+     *
+     * @return Headers
+     */
+    public function addLastUpdated(?string $last_updated = null): Headers
+    {
+        $this->headers->add('X-Last-Updated', $last_updated);
 
         return $this;
     }